remove all directories and refactored structure
